Code review comments addressed.
